% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{The length of a string}
\usage{
hello()
}
\value{
A numeric vector giving number of characters (code points) in each
element of the character vector. Missing string have missing length.
}
\description{
The length of a string
}
