% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldcm.R
\docType{package}
\name{ldcm-package}
\alias{ldcm}
\alias{ldcm-package}
\title{ldcm: Longitudinal Diagnostic Classification Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Diagnostic classification models (DCMs) can provide useful information about students' strengths and weaknesses in an academic domain beyond the norm-referenced or percentile results typically achieved through item response theory. DCMs, however, need to be able to accommodate the multilevel framework in which most education takes place, or they run the risk of providing substantially inaccurate results. The purpose of this project is to develop and make accessible to applied researchers a multilevel extension to the longitudinal diagnostic classification model (DCM) to help researchers take into account contextual effects that can impact the fidelity and effectiveness of an educational intervention.
}
\details{
See \code{vignette("ldcm", package = "ldcm")} for an overview
of the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cotterell.github.io/ldcm}
  \item \url{https://github.com/cotterell/ldcm}
  \item Report bugs at \url{https://github.com/cotterell/ldcm/issues}
}

}
\author{
\strong{Maintainer}: Michael E. Cotterell \email{mepcott@uga.edu} (\href{https://orcid.org/0000-0002-4302-0891}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Matthew Madison \email{mjmadison@uga.edu} [copyright holder]
  \item Minjeong Jeon \email{mjjeon@ucla.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item University of Georgia (The content and opinions expressed in this material do not necessarily reflect the views of nor are they endorsed by the University of Georgia or the University System of Georgia.) [copyright holder]
  \item U.S. Department of Education Institute of Education Sciences (This work is supported by the U.S. Department of Education Institute of Education Sciences under IES Award Number R305D220020. Any opinions, findings, conclusions or recommendations expressed in this material are those of the author(s) and do not necessarily reflect those of the Institute of Educational Sciences or the U.S. Department of Education.) [funder]
}

}
\keyword{internal}
